Replace substr in getCode() to cover more cases.
The current replacer does not cover all cases to remove declare statements regarding 'strict_types' definitions.

Currently not covered:
* If there is more than one line break before the declare statement
* If there are arbitrary spaces within the declare statement
* If there is a comment before the declare statement

To not build the whole statemachine by ourself I decided to use a regex to tackle those problems. As this will in most cases only be ran if the boostrap needs to be regenerated we should be safe with the implied performance hit regarding the usage of regex here.

Rundown on the Regex parts:

* `^` - Match the beginning of the file only
* `(<\?php)?` - Match starting PHP Tags (if existent)
* `([\s]*/\*\*?.*?\*/[\s]*)?` - Match file comments (if existent)
* `[\s]*` - Match arbitrary number of spaces
* `(declare[\s]*\([\s]*strict_types[\s]*=[\s]*1[\s]*\);)?` - Match strict definition (if existent)

This should handle the whole pretty cleanup more gracefully regarding external/vendored codebases which may not apply to some predefined CodingStandard.